+++ tests/util/ServicesProviderTrait.php
@@ -56,7 +56,7 @@ trait ServicesProviderTrait
     /**
      * @throws \Exception
      */
-    public function getConfiguredContainer(bool $ensureBinariesExists = false): ContainerInterface
+    public function getConfiguredContainer(bool $ensureBinariesExists = false, ?string $ffmpegBinary = null, ?string $ffprobeBinary = null): ContainerInterface
     {
         if (!defined('FFMPEG_BINARY_PATH')) {
             throw new \Exception('Missing phpunit constant FFMPEG_BINARY_PATH');
@@ -65,10 +65,11 @@ trait ServicesProviderTrait
             throw new \Exception('Missing phpunit constant FFPROBE_BINARY_PATH');
         }
 
-        $ffmpegBinary  = FFMPEG_BINARY_PATH;
-        $ffprobeBinary = FFPROBE_BINARY_PATH;
+        $ffmpegBinary  = $ffmpegBinary ?? FFMPEG_BINARY_PATH;
+        $ffprobeBinary = $ffprobeBinary ?? FFPROBE_BINARY_PATH;
+
         if ($ensureBinariesExists) {
-            if (mb_strpos(FFMPEG_BINARY_PATH, './') !== false) {
+            if (mb_strpos($ffmpegBinary, './') !== false) {
                 // relative directory
                 $ffmpegBinary = realpath(FFMPEG_BINARY_PATH);
                 if ($ffmpegBinary === false || !file_exists($ffmpegBinary)) {
@@ -80,7 +81,7 @@ trait ServicesProviderTrait
                 }
             }
 
-            if (mb_strpos(FFPROBE_BINARY_PATH, './') !== false) {
+            if (mb_strpos($ffprobeBinary, './') !== false) {
                 // relative directory
                 $ffprobeBinary = realpath(FFPROBE_BINARY_PATH);
                 if ($ffprobeBinary === false || !file_exists($ffprobeBinary)) {